backend/requirements.txt
# VetLayer Backend Dependencies
# ── Core ─────────────────────────────────────────────────────────────
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic[email]==2.10.4
pydantic-settings==2.7.1

# ── Database ─────────────────────────────────────────────────────────
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.1

# ── LLM ──────────────────────────────────────────────────────────────
openai==1.58.1
anthropic==0.42.0

# ── File parsing ─────────────────────────────────────────────────────
pypdf>=4.0.0
python-docx==1.1.2
python-multipart==0.0.18

# ── PDF generation ──────────────────────────────────────────────────
reportlab>=4.0.0

# ── Authentication ───────────────────────────────────────────────────
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.1

# ── Utilities ────────────────────────────────────────────────────────
python-dotenv==1.0.1
httpx==0.28.1

# ── Testing ──────────────────────────────────────────────────────────
pytest==8.3.4
pytest-asyncio==0.25.0
pytest-cov==6.0.0
